Fix a Safari language redirect fuckup
It seems that Safari keeps the permanent redirects in cache, so when changing your language and go to the same url without the language in the url it redirects from itself to the url with language in cache, so the cookie doesn't work. When, only for the language redirect an temprary move is used, it will work like we should expect.
This is a "bug" in all Fork 3 versions, in Fork 2 it wasn't a problem yet.